backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
asyncpg>=0.30.0
redis[hiredis]>=5.2.0
pydantic-settings>=2.6.0
typing_extensions>=4.8.0
PyJWT>=2.8.0
python-multipart>=0.0.12
slowapi>=0.1.9
